refactor(connection): split wasm implementation into a separate module
